# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de registro de asistencia",
        "description": "listado de URI's de la API alumnos",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/docentes": {
            "get": {
                "tags": [
                    "Docente"
                ],
                "summary": "lista de todos los Docentes",
                "operationId": "b85ffdc9da338fc6fdaad61f940e2aef",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Paolo Lucio"
                                                    },
                                                    "apellido": {
                                                        "type": "string",
                                                        "example": "Conza HAquehua"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "paolo@gmail.com"
                                                    },
                                                    "telefono": {
                                                        "type": "string",
                                                        "example": "987288269"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Docente"
                ],
                "summary": "Registrar a un nuevo Alumno",
                "operationId": "8dcbf1c55f5d3d4b18b8a65047973c4c",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "nombre": {
                                                "type": "string"
                                            },
                                            "apellido": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "telefono": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombre": "Paolo Lucio",
                                    "apellido": "Conza Haquehua",
                                    "email": "paolo@gmail.com",
                                    "telefono": "12345678"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Paolo Lucio"
                                        },
                                        "apellido": {
                                            "type": "string",
                                            "example": "Conza Haquehua"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "paolo@gmail.com"
                                        },
                                        "telefono": {
                                            "type": "string",
                                            "example": "12345678"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE CONTENT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The apellidos field is required."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Objeto de errores"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/docentes/{id}": {
            "put": {
                "tags": [
                    "Docente"
                ],
                "summary": "Actualizar la información de un alumno",
                "operationId": "1e0be6fde94ea3bea7aadc7f24930e64",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "nombre": {
                                                "type": "string"
                                            },
                                            "apellido": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "telefono": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombre": "Paolo Lucio Editado",
                                    "apellido": "Conza Haquehua Editado",
                                    "email": "paolo@gmail.com Editado",
                                    "telefono": "12345678 Editado"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Paolo Lucio Editado"
                                        },
                                        "apellido": {
                                            "type": "string",
                                            "example": "Conza Haquehua Editado"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "paolo@gmail.com Editado"
                                        },
                                        "telefono": {
                                            "type": "string",
                                            "example": "12345678 Editado"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE CONTENT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The apellidos field is required."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Objeto de errores"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Docente"
                ],
                "summary": "Eliminar la información de un alumno",
                "operationId": "e155943689a25a5738dc8ba0f7ad00db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "NO CONTENT"
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No se pudo realizar correctamente la operación"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}